+++ R/get_precision.R
@@ -5,8 +5,10 @@
 
 get_precision <- function(x) {
 	LEN <- max(nchar(sub(".", "", paste(x), fixed=TRUE)))
-	PRES <- max(nchar(paste(as.integer(x))))
-	PRES <- LEN - PRES
+	W <- paste(as.integer(x))
+	W[W == "NA"] <- NA
+	W <- max(nchar(W), na.rm=TRUE)
+	PRES <- LEN - W
 	cat("length:", LEN, "\n")
 	cat("precision:", PRES, "\n")
 }